✨ feat(lib): Add remove_diacritics function to strip diacritics from Arabic text 🐛 fix(lib/quran/verse.rs): Replace usage of strip_tashkeel with remove_diacritics 🔧 chore(lib/util): Change strip_tashkeel function to be private A new function called `remove_diacritics` has been added to the `util` module. This function removes diacritics from Arabic text, including characters such as vowel marks and other diacritical marks.
The `strip_tashkeel` function has been updated to be private, as it is no longer used externally. This change is to ensure encapsulation and prevent accidental misuse.

In the `quran/verse.rs` module, the usage of `strip_tashkeel` has been replaced with `remove_diacritics` to achieve the same functionality but with a more descriptive name.

Please note that the tests for the `strip_tashkeel` function have been left intact for reference and will be removed in a future commit.

Signed-off-by: Sabry Awad <dr.sabry1997@gmail.com>